My Code Review 2 Changes
