+++ app/src/main/java/edu/uc/forbesne/shoppingsidekick/dto/MarketApiObject.kt
@@ -1,6 +1,10 @@
 package edu.uc.forbesne.shoppingsidekick.dto
 
-
-// Used to store data fetched from Market API
+/**
+ * Data calss that represents Market Object
+ *
+ * @param market Market Object
+ * @param products List of all the products
+ */
 class MarketApiObject(var market: Market = Market(), var products: ArrayList<Product> = ArrayList<Product>()) {
 }
