+++ app/src/main/java/edu/uc/forbesne/shoppingsidekick/dto/MarketProductPrice.kt
@@ -1,5 +1,11 @@
 package edu.uc.forbesne.shoppingsidekick.dto
 
-// Represents a markets price for one product
+/**
+ * Data Class that represents market product price
+ *
+ * @param shopName Name of the shop
+ * @param upc UPC of the product
+ * @param price Price of the product
+ */
 data class MarketProductPrice (var shopName : String ="", var upc :String = "", var price: Float =0.0f) {
 }
