+++ app/src/main/java/edu/uc/forbesne/shoppingsidekick/dto/ProductPriceList.kt
@@ -1,7 +1,10 @@
 package edu.uc.forbesne.shoppingsidekick.dto
 
-// Used to represent one product's array of markets prices.
-// Size is hard coded to 3 in mainViewModel (for 3 APIs)
+/**
+ * Data class to store Product Price List from different markets.
+ *
+ * @param size number of markets
+ */
 data class ProductPriceList (var size :Int ) {
     val list:Array<MarketProductPrice> = Array<MarketProductPrice>(size){MarketProductPrice()}
 }
